# gpt-sovits-backend/GPT-SoVITS.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GPT-SoVITS",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/models/{model_id}": {
      "get": {
        "summary": "获取模型详情",
        "deprecated": false,
        "description": "获取模型详情",
        "operationId": "get_models_get_model_detail",
        "tags": [
          "models"
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "description": "路径参数: model_id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "summary": "更新模型信息",
        "deprecated": false,
        "description": "更新模型信息",
        "operationId": "put_models_update_model",
        "tags": [
          "models"
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "description": "路径参数: model_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "删除模型",
        "deprecated": false,
        "description": "删除模型",
        "operationId": "delete_models_delete_model",
        "tags": [
          "models"
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "description": "路径参数: model_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/models/{model_id}/toggle-public": {
      "post": {
        "summary": "切换模型公开状态",
        "deprecated": false,
        "description": "切换模型公开状态",
        "operationId": "post_models_toggle_model_public",
        "tags": [
          "models"
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "description": "路径参数: model_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/models/my-models": {
      "get": {
        "summary": "获取用户的模型列表",
        "deprecated": false,
        "description": "获取用户的模型列表",
        "operationId": "get_models_get_my_models",
        "tags": [
          "models"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "查询参数: page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "查询参数: per_page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "查询参数: status  可选值：active/inactive",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/models/stats": {
      "get": {
        "summary": "获取用户模型统计信息",
        "deprecated": false,
        "description": "获取用户模型统计信息",
        "operationId": "get_models_get_model_stats",
        "tags": [
          "models"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/models/tags": {
      "get": {
        "summary": "获取可用标签列表",
        "deprecated": false,
        "description": "获取可用标签列表",
        "operationId": "get_models_get_tags",
        "tags": [
          "models"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/tts/models/{model_id}": {
      "get": {
        "summary": "获取模型详情",
        "deprecated": false,
        "description": "获取模型详情",
        "operationId": "get_tts_get_model_detail",
        "tags": [
          "tts"
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "description": "路径参数: model_id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "headers": {}
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "examples": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}